+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.0] - 2020-12-10
+
 ### Added
 
 - Add support for setting kubeconfig secret in app CRs for remote clusters.
@@ -88,7 +90,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add initial version that implements InstallApps for use in apptestctl and
 Go integration tests.
 
-[Unreleased]: https://github.com/giantswarm/apptest/compare/v0.7.1...HEAD
+[Unreleased]: https://github.com/giantswarm/apptest/compare/v0.8.0...HEAD
+[0.8.0]: https://github.com/giantswarm/apptest/compare/v0.7.1...v0.8.0
 [0.7.1]: https://github.com/giantswarm/apptest/compare/v0.7.0...v0.7.1
 [0.7.0]: https://github.com/giantswarm/apptest/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/giantswarm/apptest/compare/v0.5.0...v0.6.0